+++ src/test/java/io/blamer/bot/conversation/routes/TokenTest.java
@@ -22,9 +22,10 @@
  * SOFTWARE.
  */
 
-package io.blamer.bot.conversation;
+package io.blamer.bot.conversation.routes;
 
 import annotation.TestWithSpringContext;
+import io.blamer.bot.conversation.routes.Token;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Assertions;
@@ -34,53 +35,35 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.commands.BotCommand;
 
 @TestWithSpringContext
 @ExtendWith(MockitoExtension.class)
-class RegistryTest {
+class TokenTest {
 
-    @Autowired
-    private Registry messageGenerator;
+  @Autowired
+  private Token tokenMessageGenerator;
 
-    @Test
-    void createsRegistryMessage(@Mock final Update update, @Mock final Message message) {
-        final String token = "tkn";
-        final String template = "Registry token doesn't available yet, sorry. Your token: '%s'";
-        Mockito.when(message.getText()).thenReturn("/registry tkn");
-        Mockito.when(update.getMessage()).thenReturn(message);
-        Assertions.assertNotNull(messageGenerator.messageFromUpdate(update));
-        MatcherAssert.assertThat(
-            "Response contains right text",
-            messageGenerator.messageFromUpdate(update).getText(),
-            Matchers.equalTo(template.formatted(token))
-        );
-    }
+  @Test
+  void createsStartMessage(@Mock final Update update, @Mock final Message message) {
+    Mockito.when(update.getMessage()).thenReturn(message);
+    Assertions.assertNotNull(this.tokenMessageGenerator.messageFromUpdate(update));
+    MatcherAssert.assertThat(
+      "Response in context right text",
+      this.tokenMessageGenerator.messageFromUpdate(update).getText(),
+      Matchers.equalTo("test token message")
+    );
+  }
 
-    @Test
-    void createsRegistryMessageWithError(@Mock final Update update, @Mock final Message message) {
-        final String expected = "Token not found in /registry";
-        Mockito.when(message.getText()).thenReturn("/registry");
-        Mockito.when(update.getMessage()).thenReturn(message);
-        final SendMessage actual = messageGenerator.messageFromUpdate(update);
-        Assertions.assertNotNull(actual);
-        MatcherAssert.assertThat(
-            "Response contains right text",
-            actual.getText(),
-            Matchers.equalTo(expected)
-        );
-    }
-
-    @Test
-    void createsBotCommand() {
-        final BotCommand actual = this.messageGenerator.messageAsBotCommand();
-        MatcherAssert.assertThat(actual.getCommand(), Matchers.equalTo("/registry"));
-        MatcherAssert.assertThat(
-            actual.getDescription(),
-            Matchers.equalTo("test registry description")
-        );
-    }
+  @Test
+  void createsDescription() {
+    final BotCommand actual = this.tokenMessageGenerator.messageAsBotCommand();
+    MatcherAssert.assertThat(actual.getCommand(), Matchers.equalTo("/token"));
+    MatcherAssert.assertThat(
+      actual.getDescription(),
+      Matchers.equalTo("test token description")
+    );
+  }
 }
